+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Added
 - Stateful testing support via ``Open API links`` syntax. `#548`_
 - New ``add_case`` hook. `#458`_
 - Support for parameters serialization formats in Open API 2 / 3. For example ``pipeDelimited`` or ``deepObject``. `#599`_
+- Support serializing parameters with ``application/json`` content type. `#594`_
 
 Changed
 ~~~~~~~
@@ -1096,6 +1097,7 @@ Fixed
 .. _#600: https://github.com/kiwicom/schemathesis/issues/600
 .. _#599: https://github.com/kiwicom/schemathesis/issues/599
 .. _#596: https://github.com/kiwicom/schemathesis/issues/596
+.. _#594: https://github.com/kiwicom/schemathesis/issues/594
 .. _#586: https://github.com/kiwicom/schemathesis/issues/586
 .. _#582: https://github.com/kiwicom/schemathesis/issues/582
 .. _#579: https://github.com/kiwicom/schemathesis/issues/579